Log master Mcp/requirements.txt
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
structlog==23.2.0
python-dateutil==2.8.2
anyio==4.2.0
jsonschema==4.20.0
pydantic==2.5.2
certifi==2023.11.17
